CH-338 Addressing review comments
