--- provisioner/config/haloprov.cfg
@@ -1,42 +1,0 @@
-# Copyright (c) 2022 Seagate Technology LLC and/or its Affiliates
-
-# This program is free software: you can redistribute it and/or modify it
-# under the terms of the GNU Affero General Public License as published by
-# the Free Software Foundation, either version 3 of the License,
-# or (at your option) any later version.
-
-# This program is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
-# See the GNU Affero General Public License for more details.
-
-# You should have received a copy of the GNU Affero General Public License
-# along with this program. If not, see <https://www.gnu.org/licenses/>.
-# For any questions about this software or licensing, please email
-# opensource@seagate.com or cortx-questions@seagate.com.
-
-[InfraMonitor]
-resource_type=SOFTWARE
-resource_id=halo_infra_monitor
-config_cmd=/opt/install_depot/infra_monitor/infra_monitor_config
-unconfig_cmd=/opt/install_depot/infra_monitor/infra_monitor_unconfig
-validate_cmd=/opt/install_depot/infra_monitor/infra_monitor_validate
-[LsiHBA]
-resource_type=SERVER_HBA
-resource_id=lsihba
-config_cmd=/opt/install_depot/lsihba/server_hba_config
-unconfig_cmd=/opt/install_depot/lsihba/server_hba_unconfig
-validate_cmd=/opt/install_depot/lsihba/server_hba_validate
-[MellanoxNIC]
-resource_type=SERVER_NIC
-resource_id=mellanoxnic
-config_cmd=/opt/install_depot/mellanoxnic/server_nic_config
-unconfig_cmd=/opt/install_depot/mellanoxnic/server_nic_unconfig
-validate_cmd=/opt/install_depot/mellanoxnic/server_nic_validate
-[Kubernetes]
-resource_type=CLUSTER
-resource_id=kubernetes
-install_cmd=/opt/install_depot/kubernetes/cluster_setup
-uninstall_cmd=/opt/install_depot/kubernetes/cluster_teardown
-validate_cmd=/opt/install_depot/kubernetes/cluster_validate
-
