+++ provisioner/config/nodeprep.yaml
@@ -1,0 +1,46 @@
+# Copyright (c) 2022 Seagate Technology LLC and/or its Affiliates
+
+# This program is free software: you can redistribute it and/or modify it
+# under the terms of the GNU Affero General Public License as published by
+# the Free Software Foundation, either version 3 of the License,
+# or (at your option) any later version.
+
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
+# See the GNU Affero General Public License for more details.
+
+# You should have received a copy of the GNU Affero General Public License
+# along with this program. If not, see <https://www.gnu.org/licenses/>.
+# For any questions about this software or licensing, please email
+# opensource@seagate.com or cortx-questions@seagate.com.
+
+Ansible:
+    resource_type: SOFTWARE
+    resource_id: Ansible
+    install_cmd: /opt/halo/install_depot/cortx-halo/provisioner/components/ansible/setup.py
+    config_cmd: /opt/halo/install_depot/cortx-halo/provisioner/components/ansible/config.py
+    validate_cmd: /opt/halo/install_depot/cortx-halo/provisioner/components/ansible/validate.py
+    teardown_cmd: /opt/halo/install_depot/cortx-halo/provisioner/components/ansible/teardown.py
+K8s_cri:
+    resource_type: SOFTWARE
+    resource_id: K8s_cri
+    install_cmd: /opt/halo/install_depot/cortx-halo/provisioner/components/k8s/setup.py
+    config_cmd: /opt/halo/install_depot/cortx-halo/provisioner/components/k8s/config.py
+    validate_cmd: /opt/halo/install_depot/cortx-halo/provisioner/components/k8s/validate.py
+    teardown_cmd: /opt/halo/install_depot/cortx-halo/provisioner/components/k8s/teardown.py
+Monitor:
+    resource_type: SOFTWARE
+    resource_id: Monitor
+    install_cmd: /opt/halo/install_depot/cortx-halo/provisioner/components/monitor/setup.py
+    config_cmd: /opt/halo/install_depot/cortx-halo/provisioner/components/monitor/config.py
+    validate_cmd: /opt/halo/install_depot/cortx-halo/provisioner/components/monitor/validate.py
+    teardown_cmd: /opt/halo/install_depot/cortx-halo/provisioner/components/monitor/teardown.py
+Nodecli:
+    resource_type: SOFTWARE
+    resource_id: K8s_cri
+    install_cmd: /opt/halo/install_depot/cortx-halo/provisioner/components/nodecli/setup.py
+    config_cmd: /opt/halo/install_depot/cortx-halo/provisioner/components/nodecli/config.py
+    validate_cmd: /opt/halo/install_depot/cortx-halo/provisioner/components/nodecli/validate.py
+    teardown_cmd: /opt/halo/install_depot/cortx-halo/provisioner/components/nodecli/teardown.py
+
