+++ provisioner/haloprov.py
@@ -17,14 +17,13 @@
 # For any questions about this software or licensing, please email
 # opensource@seagate.com or cortx-questions@seagate.com.
 
-#! /usr/bin/env python3
-
 from resource import Component
-from cmdconfig import CmdConfig
+from config import ProvisionerConfig
 from const import FileType
+import os
 
 def main():
-    cmdcfg = CmdConfig(FileType.INI,'./config/haloprov.cfg')
+    cmdcfg = ProvisionerConfig(FileType.INI, os.path.abspath('./config/haloprov.yaml'))
     resources = cmdcfg.get_resources()
     for resource in resources:
         comp = Component(resource, cmdcfg, None)
